Trent Files Thus Far Sprint 1
roomDisplay.php = a rough mock up of displaying rooms just to get them out of the db

roomRes.php = reservation form for reserving a room, needs to be linked to roomDisplay to send the room selected to this page

roomResData.php = user data is uploaded to database from roomRes as well as displays a little confirmation text.

connect,php  = connect to database function.

All of these are in rough form but i need to upload them off my laptop, most of the premise/function ill need to use is currently included in these and can be finalized once DB is created and styled later on.